Fix crash when loading GOREYE_SPAWN_EGG in creative tab
In the development environment, calling getDefaultInstance()
on GOREYE_SPAWN_EGG worked without issues. However, on normal clients this would
throw the following error when loading the creative tab icon:

    java.lang.NoSuchMethodError: 'net.minecraft.world.item.ItemStack dev.architectury.core.item.ArchitecturySpawnEggItem.getDefaultInstance()'

ArchitecturySpawnEggItem does not implement getDefaultInstance() in this
context. Changing the declared type to the base Item ensures the method is
available and prevents the crash.